requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pandas==2.2.2
google-generativeai==0.8.1
plotly==5.24.0